Create removing_correlations
This code is written in Visual basic for Applications. It should act as a macro of Excell. 
A series of random values is first  generated to simulate a 3-dimensional distribution. The software computes the 3 two-dimensional marginal distributions. According to IPF with uniform seed on one hand and maximizing the entropy (uncertainty) of the distribution on the other hand, a new uncorrelated distribution is constructed. The 2 algorithms are compared.